pyproject.toml
[tool.poetry]
name = "DjangoPro"
version = "0.1.0"
description = ""
authors = ["kskibun <costam@costam.pl>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
django-filter = "^24.3"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
